requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
awscli==1.11.127
Babel==2.4.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.23
boto==2.48.0
boto3==1.4.4
botocore==1.5.90
celery==3.1.25
certifi==2017.4.17
colorama==0.3.7
coreapi==2.3.0
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.1
django-celery==3.2.1
django-cors-headers==2.0.2
django-rest-swagger==2.1.2
django-storages==1.6.4
django-timezone-field==2.0
djangorestframework==3.6.3
docutils==0.13.1
flower==0.9.1
futures==3.1.1
gunicorn==19.7.1
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.3
json2html==1.1.1
kombu==3.0.37
librabbitmq==1.6.1
Markdown==2.6.8
MarkupSafe==1.0
oauthlib==2.0.2
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.4
pyasn1==0.3.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-memcached==1.58
pytz==2017.2
PyYAML==3.12
requests==2.14.2
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.10
simplejson==3.10.0
six==1.10.0
tornado==4.2
uritemplate==3.0.0
vine==1.1.3
whitenoise==3.3.0
